PART 1: Change temperature type to int

+++ src/main/java/icd0004/report/CurrentWeatherReport.java
@@ -5,7 +5,7 @@ import lombok.Data;
 @Data
 public class CurrentWeatherReport {
     private String date;
-    private Double temperature;
+    private int temperature;
     private int humidity;
     private int pressure;
 }
